backend/pyproject.toml
[project]
name = "who-was-i-backend"
version = "0.1.0"
description = "Ephemeral room-based chat backend"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",

  # WebSockets
  "websockets>=12.0",

  # Database (Postgres, async)
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",

  # Config & validation
  "pydantic>=2.6",
  "pydantic-settings>=2.2",

  # Utilities
  "python-dotenv>=1.0",
  "argon2-cffi>=21.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.3"
]

[tool.uvicorn]
factory = true

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*"]
